+++ tests/TestCase.php
@@ -2,29 +2,42 @@
 
 namespace Sfneal\Address\Tests;
 
+use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
 use Sfneal\Address\Models\Address;
 use Sfneal\Address\Providers\AddressServiceProvider;
-use Sfneal\Address\Tests\Models\People;
-use Sfneal\Address\Tests\Providers\TestingServiceProvider;
+use Sfneal\Testing\Models\People;
+use Sfneal\Testing\Providers\MockModelsServiceProvider;
 
 class TestCase extends OrchestraTestCase
 {
     use RefreshDatabase;
 
+    /**
+     * Register package service providers.
+     *
+     * @param Application $app
+     * @return array
+     */
     protected function getPackageProviders($app)
     {
         return [
-            TestingServiceProvider::class,
+            MockModelsServiceProvider::class,
             AddressServiceProvider::class,
         ];
     }
 
+    /**
+     * Define environment setup.
+     *
+     * @param Application $app
+     * @return void
+     */
     protected function getEnvironmentSetUp($app)
     {
         // Migrate 'people' table
-        include_once __DIR__.'/migrations/create_people_table.php.stub';
+        include_once __DIR__.'/../vendor/sfneal/mock-models/database/migrations/create_people_table.php.stub';
         (new \CreatePeopleTable())->up();
 
         // Migrate address table